Fix content length clamping issue in frontend (#1915)
Netty has two  content length related helper methods:
  public static long getContentLength(HttpMessage message, long defaultValue)
  public static int getContentLength(HttpMessage message, int defaultValue)
We were calling getContentLength(message, -1) instead of getContentLength(message, -1L)
in a couple places, which lead to the length used to tell when the request is fully sent
out getting clamped to an int, which made requests pause after 2 GB of
data was sent.

After fixing this, the download  of greater than 2 GB objects with the content-length
response headers works. This bug did not affect
transfer-encoding:chunked serving since a different end condition for
telling when the response is fully sent is used there.

+++ ambry-api/src/main/java/com/github/ambry/config/FrontendConfig.java
@@ -127,7 +127,7 @@ public class FrontendConfig {
    */
   @Config("frontend.chunked.get.response.threshold.in.bytes")
   @Default("8192")
-  public final Integer chunkedGetResponseThresholdInBytes;
+  public final long chunkedGetResponseThresholdInBytes;
 
   /**
    * Boolean indicator to specify if frontend should allow the post requests that carry serviceId used as target
@@ -258,7 +258,7 @@ public class FrontendConfig {
     pathPrefixesToRemove = Collections.unmodifiableList(
         pathPrefixesFromConfig.stream().map(this::stripLeadingAndTrailingSlash).collect(Collectors.toList()));
     chunkedGetResponseThresholdInBytes =
-        verifiableProperties.getInt("frontend.chunked.get.response.threshold.in.bytes", 8192);
+        verifiableProperties.getLong("frontend.chunked.get.response.threshold.in.bytes", 8192);
     allowServiceIdBasedPostRequest =
         verifiableProperties.getBoolean("frontend.allow.service.id.based.post.request", true);
     attachTrackingInfo = verifiableProperties.getBoolean("frontend.attach.tracking.info", true);
